fix: Adapt to Nautilus 49 API changes
Nautilus 49 beta bumps the `gobject-introspection` API version from 4.0 to 4.1, which breaks compatibility with this extension.

This commit updates the code to be compatible with the new API version.

Close: #242

Note: A complete fix for a related bug is currently blocked by an upstream issue in the nautilus-python library.

See: https://gitlab.gnome.org/GNOME/nautilus-python/-/issues/38